+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added version numbers for all tools
+
 ### Fixed
 
 ### Changed
+
+- `--version` subcommand now also shows the versions of vesyla tools
 
 ### Removed